Added compatibility with Dojo i18n
There are some unique circumstances where a file bundled by webpack may be required by something external to it, like requirejs or dojo. For Dojo, in particular, it looks for this structure on the module is returned. If this doesn't exist, it will fail to translate. This should not impact this loader in anyway shape or form. It is a quiet addition to provide backwards compatibility with annoying systems